Reuse one sort function for file related quick open results
