+++ src/vs/workbench/browser/parts/quickopen/editorHistoryModel.ts
@@ -245,30 +245,7 @@ export class EditorHistoryModel extends QuickOpenModel {
 		if (searchValue) {
 			let normalizedSearchValue = strings.stripWildcards(searchValue.toLowerCase());
 
-			return results.sort((elementA: EditorHistoryEntry, elementB: EditorHistoryEntry) => {
-
-				// Give matches with label highlights higher priority over
-				// those with only description highlights
-				const labelHighlightsA = elementA.getHighlights()[0] || [];
-				const labelHighlightsB = elementB.getHighlights()[0] || [];
-
-				if (labelHighlightsA.length && !labelHighlightsB.length) {
-					return -1;
-				} else if (!labelHighlightsA.length && labelHighlightsB.length) {
-					return 1;
-				}
-
-				// Sort by name/path
-				let nameA = elementA.getInput().getName();
-				let nameB = elementB.getInput().getName();
-
-				if (nameA === nameB) {
-					nameA = elementA.getResource().fsPath;
-					nameB = elementB.getResource().fsPath;
-				}
-
-				return comparers.compareAnything(nameA, nameB, normalizedSearchValue);
-			});
+			return results.sort((elementA: EditorHistoryEntry, elementB: EditorHistoryEntry) => QuickOpenEntry.compare(elementA, elementB, normalizedSearchValue));
 		}
 
 		// Leave default "most recently used" order if user is not actually searching
